update: manage topping, size features
