+++ coffee_shop_ad/lib/routing/app_router.dart
@@ -1,10 +1,10 @@
 import 'package:coffee_shop_admin/screens/auth/auth_screen.dart';
-import 'package:coffee_shop_admin/screens/drink_management/size_create.dart';
-import 'package:coffee_shop_admin/screens/drink_management/size_detail.dart';
-import 'package:coffee_shop_admin/screens/drink_management/size_edit.dart';
-import 'package:coffee_shop_admin/screens/drink_management/topping_create.dart';
-import 'package:coffee_shop_admin/screens/drink_management/topping_detail.dart';
-import 'package:coffee_shop_admin/screens/drink_management/topping_edit.dart';
+import 'package:coffee_shop_admin/screens/drink_manage/size_create.dart';
+import 'package:coffee_shop_admin/screens/drink_manage/size_detail.dart';
+import 'package:coffee_shop_admin/screens/drink_manage/size_edit.dart';
+import 'package:coffee_shop_admin/screens/drink_manage/topping_create.dart';
+import 'package:coffee_shop_admin/screens/drink_manage/topping_detail.dart';
+import 'package:coffee_shop_admin/screens/drink_manage/topping_edit.dart';
 import 'package:coffee_shop_admin/screens/main_page.dart';
 import 'package:coffee_shop_admin/screens/product_detail.dart';
 import 'package:coffee_shop_admin/screens/profile/profile_screen.dart';
