+++ coffee_shop_ad/lib/screens/drink_manage/size_detail.dart
@@ -1,28 +1,28 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:coffee_shop_admin/screens/drink_management/topping_card.dart';
-import 'package:coffee_shop_admin/screens/drink_management/topping_edit.dart';
-import 'package:coffee_shop_admin/services/blocs/topping_list/topping_list_bloc.dart';
-import 'package:coffee_shop_admin/services/blocs/topping_list/topping_list_event.dart';
-import 'package:coffee_shop_admin/services/models/topping.dart';
+import 'package:coffee_shop_admin/screens/drink_manage/size_edit.dart';
+import 'package:coffee_shop_admin/services/blocs/size_manage/size_list_bloc.dart';
+import 'package:coffee_shop_admin/services/blocs/size_manage/size_list_event.dart';
+import 'package:coffee_shop_admin/services/models/size.dart';
 import 'package:coffee_shop_admin/utils/colors/app_colors.dart';
 import 'package:coffee_shop_admin/utils/constants/dimension.dart';
 import 'package:coffee_shop_admin/utils/styles/app_texts.dart';
 import 'package:coffee_shop_admin/utils/styles/button.dart';
+import 'package:coffee_shop_admin/widgets/feature/drink_manage/size/size_card.dart';
 import 'package:coffee_shop_admin/widgets/global/custom_app_bar.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:quickalert/quickalert.dart';
 
-class ToppingDetail extends StatefulWidget {
-  final Topping product;
-  const ToppingDetail({super.key, required this.product});
+class SizeDetail extends StatefulWidget {
+  final Size product;
+  const SizeDetail({super.key, required this.product});
 
   @override
-  State<ToppingDetail> createState() => _ToppingDetailState();
+  State<SizeDetail> createState() => _SizeDetailState();
 }
 
-class _ToppingDetailState extends State<ToppingDetail> {
+class _SizeDetailState extends State<SizeDetail> {
   @override
   void initState() {
     super.initState();
@@ -38,13 +38,13 @@ class _ToppingDetailState extends State<ToppingDetail> {
     final isKeyboard = MediaQuery.of(context).viewInsets.bottom != 0;
 
     // ignore: no_leading_underscores_for_local_identifiers
-    void _handleOnTapEditTopping() async {
+    void _handleOnTapEditSize() async {
       Navigator.of(context)
-          .pushNamed(EditToppingScreen.routeName, arguments: widget.product);
+          .pushNamed(EditSizeScreen.routeName, arguments: widget.product);
     }
 
     // ignore: no_leading_underscores_for_local_identifiers
-    void _handleOnTapDeleteTopping() async {
+    void _handleOnTapDeleteSize() async {
       await QuickAlert.show(
         context: context,
         type: QuickAlertType.confirm,
@@ -65,13 +65,13 @@ class _ToppingDetailState extends State<ToppingDetail> {
 
           try {
             await FirebaseFirestore.instance
-                .collection("Topping")
+                .collection("Size")
                 .doc(widget.product.id)
                 .delete()
                 .then((value) {
               Navigator.of(context).pop();
               Navigator.of(context).pop();
-              BlocProvider.of<ToppingListBloc>(context).add(FetchData());
+              BlocProvider.of<SizeListBloc>(context).add(FetchData());
               FirebaseStorage.instance.refFromURL(imgUrl).delete();
               QuickAlert.show(
                 context: context,
@@ -81,7 +81,7 @@ class _ToppingDetailState extends State<ToppingDetail> {
               );
             });
           } catch (e) {
-            print("Something wrong when delete topping");
+            print("Something wrong when delete size");
             print(e);
           }
         },
@@ -101,7 +101,7 @@ class _ToppingDetailState extends State<ToppingDetail> {
                 children: [
                   CustomAppBar(
                     leading: Text(
-                      'Topping: ${widget.product.name}',
+                      'Size: ${widget.product.name}',
                       style: AppText.style.regularBlack16,
                     ),
                   ),
@@ -113,7 +113,7 @@ class _ToppingDetailState extends State<ToppingDetail> {
                       child: SingleChildScrollView(
                           child: Column(
                     children: [
-                      ToppingCard(product: widget.product),
+                      SizeCard(product: widget.product),
                       SizedBox(height: 16)
                     ],
                   ))),
@@ -147,7 +147,7 @@ class _ToppingDetailState extends State<ToppingDetail> {
                                   width: 140,
                                   child: ElevatedButton.icon(
                                       style: roundedButton,
-                                      onPressed: _handleOnTapEditTopping,
+                                      onPressed: _handleOnTapEditSize,
                                       icon: Icon(
                                         Icons.create_sharp,
                                         size: 28,
@@ -171,7 +171,7 @@ class _ToppingDetailState extends State<ToppingDetail> {
                                         Icons.delete_forever,
                                         size: 28,
                                       ),
-                                      onPressed: _handleOnTapDeleteTopping,
+                                      onPressed: _handleOnTapDeleteSize,
                                       label: Text(
                                         'Delete',
                                         style: AppText.style.regularWhite16,
