+++ coffee_shop_ad/lib/widgets/feature/drink_manage/size/size_card.dart
@@ -2,9 +2,9 @@ import 'package:coffee_shop_admin/services/models/size.dart';
 import 'package:flutter/material.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 
-import '../../../services/functions/money_transfer.dart';
-import '../../../utils/constants/dimension.dart';
-import '../../../utils/styles/app_texts.dart';
+import 'package:coffee_shop_admin/services/functions/money_transfer.dart';
+import 'package:coffee_shop_admin/utils/constants/dimension.dart';
+import 'package:coffee_shop_admin/utils/styles/app_texts.dart';
 
 class SizeCard extends StatefulWidget {
   const SizeCard({super.key, required this.product});
